The Great Merge
CreateStateVariable and EditStateVariable now use the same component, StateVariableForm, when creating/editing variables. This is pretty much a bunch of code that originally in CreateStateVariable, moved to a new component, so that EditStateVariable could use it as well
