+++ frontend/src/components/StateVariables/StateTypes.js
@@ -1,7 +1,18 @@
-const StateTypes = {
+export const StateTypes = {
   STRING: "string",
   NUMBER: "number",
   BOOLEAN: "boolean",
 };
 
-export default StateTypes;
+export const getDefaultValue = (type) => {
+  switch (type) {
+    case StateTypes.STRING:
+      return "";
+    case StateTypes.NUMBER:
+      return 0;
+    case StateTypes.BOOLEAN:
+      return false;
+    default:
+      return "";
+  }
+};
